requirements.txt
# MedGemma Trauma Analysis - Requirements

## Core
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.21.0
Pillow>=10.0.0

## Medical Imaging
pydicom>=2.4.0
SimpleITK>=2.2.0
nibabel>=5.0.0

## ML/Segmentation
segmentation-models-pytorch>=0.3.0
albumentations>=1.3.0

## Data
pandas>=2.0.0
datasets>=2.14.0,<3.0.0

## Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

## Utils
tqdm>=4.65.0
PyYAML>=6.0
python-dotenv>=1.0.0

## Web
Flask>=3.0.0
Werkzeug>=3.0.0
flask-cors>=4.0.0

## HuggingFace Ecosystem (HAI-DEF models)
transformers>=4.47.0
accelerate>=0.27.0
huggingface_hub>=0.20.0

## Quantization (for 4-bit inference on GPU / Jetson)
bitsandbytes>=0.43.0

## LoRA Fine-Tuning
peft>=0.10.0
trl>=0.8.0

## Evaluation
scikit-learn>=1.3.0
evaluate>=0.4.0
bert-score>=0.3.13

## Colab / Demo
pyngrok>=7.0.0
